+++ src/modules/identity/apis/logout.api.ts
@@ -5,7 +5,6 @@ import { Endpoint, HttpPost, Module } from '../../../../lib';
 export class LogoutApi extends Endpoint {
   main() {
     this.request.session.userId = undefined;
-    this.request.session.permissions = undefined;
     return { ok: true };
   }
 }